+++ parser/src/loader/file-loader-web.ts
