--- parser/src/parser/pdf-plan-constants.ts
@@ -1,9 +1,0 @@
-export const XSTART = 1.95;
-export const YSTART = 10.27;
-export const DAY_WIDTH = 28.36;
-export const DAY_HEIGHT = 4.687;
-
-export const DATE_X_MIN = 77;
-export const DATE_X_MAX = 100;
-export const DATE_Y_MIN = 6.8;
-export const DATE_Y_MAX = 7;
